+++ test/cuerdo/arazzo_case_test.exs
@@ -6,6 +6,30 @@ defmodule Cuerdo.ArazzoCaseTest do
   import Cuerdo.ArazzoFixtures
 
   describe "run_all/3" do
+    test "executes num_cases when halt_on_error is false" do
+      Req.Test.expect(Cuerdo.Client, 2, fn conn ->
+        %{"min_age" => min_age, "name" => name} = conn.params
+        min_age = String.to_integer(min_age)
+
+        Req.Test.json(conn, [
+          %{"name" => "#{name}abc", "age" => min_age - 2}
+        ])
+      end)
+
+      document = people_document(with_self: true)
+      workflow_id = "getPeople"
+
+      opts = [
+        json_schema_resolver: Cuerdo.Resolver,
+        num_runs: 2,
+        transform_inputs: %{},
+        halt_on_error: false
+      ]
+
+      result = Cuerdo.ArazzoCase.run_all(workflow_id, document, opts)
+      assert [%Result{status: :failed}, %Result{status: :failed}] = result
+    end
+
     test "returns result struct with failure and error cases" do
       # 1st case pass, 2nd fails
       Req.Test.expect(Cuerdo.Client, 1, fn conn ->
@@ -29,7 +53,13 @@ defmodule Cuerdo.ArazzoCaseTest do
 
       document = people_document(with_self: true)
       workflow_id = "getPeople"
-      opts = [json_schema_resolver: Cuerdo.Resolver, max_runs: 2, transform_inputs: %{}]
+
+      opts = [
+        json_schema_resolver: Cuerdo.Resolver,
+        num_runs: 3,
+        transform_inputs: %{},
+        halt_on_error: true
+      ]
 
       result = Cuerdo.ArazzoCase.run_all(workflow_id, document, opts)
       [%Result{status: :passed}, %Result{status: :failed} = error_result] = result